+++ src/Bridge/Qdrant/CHANGELOG.md
@@ -1,6 +1,12 @@
 CHANGELOG
 =========
 
+0.6
+---
+
+ * [BC BREAK] The `endpointUrl` parameter for `Store` has been removed
+ * [BC BREAK] The `apiKey` parameter for `Store` has been removed
+
 0.1
 ---